test(unit): convert test for the experimental release workflow scaffolder to vitest
for #458
